% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_tools.R
\name{bedtools}
\alias{bedtools}
\alias{set_bedtools_path}
\alias{get_bedtools_path}
\title{bedtools: utilities for genomic regions}
\usage{
set_bedtools_path(path)

get_bedtools_path()
}
\arguments{
\item{path}{Path to the bedtools executable.}
}
\value{
TRUE on successful execution, FALSE otherwise.
}
\description{
bedtools: utilities for genomic regions
}
\section{set_bedtools_path}{

}

\section{get_bedtools_path}{

}

