% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_genome.R
\name{set_binGenome_path}
\alias{set_binGenome_path}
\alias{bin_genome}
\title{Wrapper for binGenome module}
\usage{
set_binGenome_path(path)

bin_genome(
  input_dir,
  strand,
  annotation,
  output_dir,
  bins = NULL,
  quantiles = NULL,
  grep_pattern = NULL,
  fixedBinSizeDownstream = NULL,
  fixedBinSizeUpstream = NULL,
  bedgraphNames = NULL,
  annotationNames = NULL,
  cores = NULL,
  normalize = FALSE
)
}
\arguments{
\item{path}{Path to the binGenome.sh executable.}

\item{input_dir}{Path to the folder containing bedgraph files.
Only the files with .bedgraph extension are considered.}

\item{strand}{Filters files according to their strandedness.
It applies grep and filters files having "pos"/"neg" anywhere in their names.}

\item{annotation}{Annotation file (.bed)}

\item{output_dir}{Output folder for binGenome.sh}

\item{grep_pattern}{\link{optional} Filter for a grep pattern in files.}

\item{fixedBinSizeDownstream, fixedBinSizeUpstream}{\link{optional} Specifies the parameters that will be
used for binning the downstream or upstream regions, respectively. Format: "binsize:binnumber".}

\item{bedgraphNames, annotationNames}{\link{optional}}

\item{cores}{\link{optional} Number of threads to use. Defualt: max(1, parallel::detectCores() - 2)}

\item{normalize}{\link{optional}}

\item{bins/quantiles}{bins or quantiles for binGenome}
}
\value{
TRUE on successful execution, FALSE otherwise.
}
\description{
Runs the binGenome.sh on the given file(s).
Prints the log of how many regions were too small for binning in each sample.
If stranded, pos and neg are merged and stripped from the name.
}
\section{set_binGenome_path}{

}

