% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bt_tools.R
\name{bt_genomecov}
\alias{bt_genomecov}
\title{Get the genome coverage using bedtools}
\usage{
bt_genomecov(input_bam, output_file = NULL)
}
\arguments{
\item{output_file}{Optional path to output file.}

\item{input_bed}{Path to input BED file.}

\item{genome_file}{Path to genome file.}
}
\value{
If output file path given, a string of its path is returned,
otherwise a data table of the result captured from the command is returned.
}
\description{
Handles following command: bedtools genomecov -ibam input_file -bg > output_file.
Used for its side effect, creating coverage file.
}
\examples{
\dontrun{
  set_bedtools_path("/usr/bin/bedtools")
  bt_genomecov("bamfile.bam", "output.bedgraph")
}
}
