% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{setConfig}
\alias{setConfig}
\alias{setLog}
\title{Set config and log file paths for makeRegions}
\usage{
setConfig(file)

setLog(file)
}
\arguments{
\item{file}{Character string specifying the file path. The containing
directory will be created if it does not exist. Use \code{NULL} to unset.}
}
\value{
No return value, called for side effects.
}
\description{
These functions set and create the files for config (.json) and log (.log)
files used by the \code{makeRegions} process.
}
\details{
\itemize{
\item \code{setConfig()} sets the file path for the configuration file.
\item \code{setLog()} sets the file path for the log file.
}
}
\examples{
\dontrun{
setConfig("~/project/config.json")
setLog("~/project/console.log")
}
}
