% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{name_output_with}
\alias{name_output_with}
\title{Utility function: output nomenclature}
\usage{
name_output_with(enclosing_dir, pars, ext)
}
\arguments{
\item{enclosing_dir}{Indicates in which dir the file will should be.}

\item{pars}{A named list of arguments. The names correspond to the parameters of the values.}

\item{ext}{Extension.}
}
\value{
A character vector as file path generated by the parameters.
}
\description{
Creates a standard name based on given parameters.
Originally created for use with make_windows, hence parameters like
input_file, upstream and downstream will be ignored.
}
\examples{
pars$make_windows$path_to_output <- name_output_with(out_dir, pars$make_windows, ".bed")
}
