% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_metagene_experiment}
\alias{plot_metagene_experiment}
\title{Pairwise and all}
\usage{
plot_metagene_experiment(
  plot_dir,
  run_dir,
  annotation_name,
  title = annotation_name,
  condition_mapper,
  aggregate_FUN = mean,
  threads = 1
)
}
\arguments{
\item{plot_dir}{Where to save PDF outputs of plotting}

\item{annotation_name}{Indicates the annotation used for binning. Appends to the file name.
If a named vector, the name is used for the title!}

\item{condition_mapper}{A data frame with columns "grep_name", "color", "display_name"
to match conditions with line colors and names to display in the legend.}

\item{threads}{Number of threads to use in pairwise plotting.
Usage makes sense if multiple pairs are plotted, e.g. nrow(condition_mapper) > 1,
and cores are available, e.g. if detectCores() >= threads or
upstream threading has not used them up. Since this utilizes mclapply from
package parallel, this option should be used careful with nested parallelization.}

\item{experiment_dir}{The run folder where windows and binning took place.
The folder should have a 'config.json' file where the parameters for make_windows (optional)
and bin_genome (required) are given.}
}
\description{
Plots all pairwise comparisons and one with all conditions.
Exploits multithreading from package parallel using mclapply on pairs generated.
binMatrix objects in the coverage folder in run_dir are aggregated
using aggregate_FUN (default: mean) and saved as rds objects in a directory
called 'aggregated_binmatrices' as a cache for further calls of the same function
with same run_dir, annotation and aggregate function.
}
