% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotting}
\alias{plotting}
\alias{source_plotting_lib}
\alias{draw_metagene}
\title{Wrapper for sourcing plotting_lib.R}
\usage{
source_plotting_lib(path)

draw_metagene(
  plot_dir,
  file_name,
  binMatrixList,
  pvThresholdsVec = c(0.05, 10^-3, 10^-5),
  ...
)
}
\arguments{
\item{path}{Path to the plotting_lib.R}

\item{plot_dir}{Where to save PDF of the plot}

\item{file_name}{File identifier. "metagene_" gets added as prefix.}

\item{binMatrixList}{List of binMatrix instances.}

\item{pvThresholdsVec}{A vector of numeric between 0 and 1 to use
as p-value thresholds in defaultPvalueColorTransformer for Wilcoxon test.}
}
\value{
Normalized path.
}
\description{
Basic plot drawing function wrapped from binGenome plotting library
}
