% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_genome.R
\name{bin_genome}
\alias{bin_genome}
\alias{set_binGenome_path}
\title{Wrapper for binGenome module}
\usage{
set_binGenome_path(path)

bin_genome(
  input_dir,
  strand,
  annotation,
  output_dir,
  strand_pattern = "pos/neg",
  bins = NULL,
  quantiles = NULL,
  grep_pattern = NULL,
  fixedBinSizeDownstream = NULL,
  fixedBinSizeUpstream = NULL,
  bedgraphNames = NULL,
  annotationNames = NULL,
  cores = NULL,
  normalize = FALSE,
  tmpDir = NULL
)
}
\arguments{
\item{path}{Path to the binGenome.sh executable.}

\item{input_dir}{Path to the folder containing bedgraph files.
Only the files with .bedgraph extension are considered.}

\item{strand}{Filters annotation files according to their strandedness.
It applies grep and filters files having "pos"/"neg" anywhere in their names.}

\item{annotation}{Annotation file (.bed)}

\item{output_dir}{Output folder for binGenome.sh}

\item{strand_pattern}{Identifiers for positive and negative strand to look in the names
of annotation files, separated by '/', first corresponding to positive
and second to negative. Default is "pos/neg".
The pattern is used literally, so user should not escape regex metacharacters.
For example, by strand_pattern = "+/-", plus is matched literally without any problem.}

\item{grep_pattern}{\link{optional} Filter for a grep pattern in files. perl is TRUE.}

\item{fixedBinSizeDownstream, fixedBinSizeUpstream}{\link{optional} Specifies the parameters that will be
used for binning the downstream or upstream regions, respectively. Format: "binsize:binnumber".}

\item{bedgraphNames, annotationNames}{\link{optional}}

\item{cores}{\link{optional} Number of threads to use. Defualt: max(1, parallel::detectCores() - 2)}

\item{normalize}{\link{optional}}

\item{bins/quantiles}{bins or quantiles for binGenome}
}
\value{
TRUE on successful execution, FALSE otherwise.
}
\description{
Runs the binGenome.sh on the given file(s).
Prints the log of how many regions were too small for binning in each sample.
If stranded, pos and neg are merged and stripped from the name.
}
