% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_windows.R
\name{make_windows}
\alias{make_windows}
\title{Create genomic regions with upstream-downstream sizes and optional filters}
\usage{
make_windows(
  input_file,
  upstream,
  downstream,
  path_to_output = NULL,
  feature = NULL,
  biotype = NULL,
  position = NULL
)
}
\arguments{
\item{input_file}{GTF or BED file}

\item{upstream}{Upstream size}

\item{downstream}{Downstream size}

\item{path_to_output}{(Optional) File path to write output. Creates dirs if they do not exist}

\item{feature}{(Optional) Filter by given feature such as gene, transcript...}

\item{biotype}{(Optional) Filter by given biotype, if exists in the input GTF, such as lncRNA, pseudogene...}

\item{position}{(Optional) Specify 'start' or 'end' to build the window around.
If not specified, the start and end coordinates are extended with upstream and downstream sized in corresponding way.}
}
\value{
A data.frame in .bed format with specified window and filter criteria.
The columns 'seqnames' and 'strand' are factors.
}
\description{
Create genomic regions with upstream-downstream sizes and optional filters
}
\examples{
example_file <- system.file("extdata", "data", "example.gtf", package = "makeRegions")
result <- make_windows(input_file=example_file,
                       upstream = 1000, downstream = 2000,
                       feature = "gene", biotype = "lncRNA",
                       position = "start")$result
print(head(result))
}
