% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_windows.R
\name{make_windows}
\alias{make_windows}
\title{Create genomic regions with with given upstream and downstream sizes}
\usage{
make_windows(
  input_file,
  upstream,
  downstream,
  path_to_output = NULL,
  feature = NULL,
  biotype = NULL,
  position = NULL
)
}
\arguments{
\item{upstream}{Upstream size}

\item{downstream}{Downstream size}

\item{path_to_output}{(Optional) File path to write output. Creates dirs if they do not exist}

\item{feature}{(Optional) Filter by given feature such as gene, transcript...}

\item{biotype}{(Optional) Filter by given biotype, if exists in the input GTF, such as lncRNA, pseudogene...}

\item{position}{(Optional) Specify 'start' or 'end' to build the window around.
If not specified, the start and end coordinates are extended with upstream and downstream sized in corresponding way.}

\item{input}{GTF or BED file}
}
\value{
Annotated .bed file with windows specified.
}
\description{
Create genomic regions with with given upstream and downstream sizes
}
\examples{
TODO
}
